* apt-pkg/acquire-item.cc (pkgAcqMetaSig::pkgAcqMetaSig): Don't 	move old version into partial/, we need it as a backup in the case 	the acquiring fails. 	(pkgAcqMetaSig::Custom600Headers): Updated to find the old 	version. 	(pkgAcqMetaSig::Done): Make sure that the file is in partial/ and 	no longer in lists/. 	(pkgAcqMetaSig::Failed): Use the old, existing version for a 	transient acquire failure. 	(pkgAcqMetaIndex::Failed): Likewise, start the gpgv pass for a 	transient failure when we have an old version and a signature 	file.
git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk@5271 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ apt/apt-pkg/acquire-item.cc
@@ -332,23 +332,6 @@ pkgAcqMetaSig::pkgAcqMetaSig(pkgAcquire *Owner,
    Desc.ShortDesc = ShortDesc;
    Desc.URI = URI;
    
-      
-   string Final = _config->FindDir("Dir::State::lists");
-   Final += URItoFileName(RealURI);
-   struct stat Buf;
-   if (stat(Final.c_str(),&Buf) == 0)
-   {
-      // File was already in place.  It needs to be re-verified
-      // because Release might have changed, so Move it into partial
-      Rename(Final,DestFile);
-      // unlink the file and do not try to use I-M-S and Last-Modified
-      // if the users proxy is broken
-      if(_config->FindB("Acquire::BrokenProxy", false) == true) {
-	 std::cerr << "forcing re-get of the signature file as requested" << std::endl;
-	 unlink(DestFile.c_str());
-      }
-   }
-
    QueueURI(Desc);
 }
 									/*}}}*/
@@ -357,8 +340,10 @@ pkgAcqMetaSig::pkgAcqMetaSig(pkgAcquire *Owner,
 /* The only header we use is the last-modified header. */
 string pkgAcqMetaSig::Custom600Headers()
 {
-   struct stat Buf;
-   if (stat(DestFile.c_str(),&Buf) != 0)
+  string Final = _config->FindDir("Dir::State::lists");
+  Final += URItoFileName(RealURI);
+  struct stat Buf;
+   if (stat(Final.c_str(),&Buf) != 0)
       return "\nIndex-File: true";
 
    return "\nIndex-File: true\nLast-Modified: " + TimeRFC1123(Buf.st_mtime);
@@ -388,6 +373,21 @@ void pkgAcqMetaSig::Done(string Message,unsigned long Size,string MD5,
 
    Complete = true;
 
+   string Final = _config->FindDir("Dir::State::lists");
+   Final += URItoFileName(RealURI);
+   if (StringToBool(LookupTag(Message,"IMS-Hit"),false))
+     {
+       // Move it into position
+       Rename (Final, DestFile);
+     }
+   else
+     {
+       // Delete the old version in lists/.  The new version will be moved
+       // there from partial/ when the signature verification succeeds
+       //
+       unlink (Final.c_str ());
+     }
+
    // queue a pkgAcqMetaIndex to be verified against the sig we just retrieved
    new pkgAcqMetaIndex(Owner, MetaIndexURI, MetaIndexURIDesc, MetaIndexShortDesc,
 		       DestFile, IndexTargets, MetaIndexParser);
@@ -396,31 +396,61 @@ void pkgAcqMetaSig::Done(string Message,unsigned long Size,string MD5,
 									/*}}}*/
 void pkgAcqMetaSig::Failed(string Message,pkgAcquire::MethodConfig *Cnf)
 {
-   // Delete any existing sigfile, so that this source isn't
-   // mistakenly trusted
-   string Final = _config->FindDir("Dir::State::lists") + URItoFileName(RealURI);
-   unlink(Final.c_str());
+   string Final =
+     _config->FindDir("Dir::State::lists") + URItoFileName(RealURI);
 
-   // if we get a timeout if fail
-   if(LookupTag(Message,"FailReason") == "Timeout" || 
-      LookupTag(Message,"FailReason") == "TmpResolveFailure") {
-      Item::Failed(Message,Cnf);
-      return;
-   }
+   // If this is a transient failure, we use the old version, if we
+   // have one.  "Using it" means moving it into the partial/
+   // directory for further consumption by gpgv.
+   //
+   if (StringToBool(LookupTag(Message,"Transient-Failure"),false))
+     {
+       cerr << "transient " << Final << "\n";
+       struct stat Buf;
+       if (stat(Final.c_str(),&Buf) == 0)
+	 {
+	   cerr << "have old version\n";
+	   Rename (Final, DestFile); 
+	   new pkgAcqMetaIndex(Owner, MetaIndexURI, MetaIndexURIDesc,
+			       MetaIndexShortDesc,
+			       DestFile, IndexTargets, MetaIndexParser);
+	 }
+       else
+	 {
+	   // queue a pkgAcqMetaIndex with no sigfile
+	   new pkgAcqMetaIndex(Owner, MetaIndexURI, MetaIndexURIDesc,
+			       MetaIndexShortDesc,
+			       "", IndexTargets, MetaIndexParser);
+	 }
+     }
+   else
+     {
+       // Delete any existing sigfile, so that this source isn't
+       // mistakenly trusted
+       unlink(Final.c_str());
 
-   // queue a pkgAcqMetaIndex with no sigfile
-   new pkgAcqMetaIndex(Owner, MetaIndexURI, MetaIndexURIDesc, MetaIndexShortDesc,
-		       "", IndexTargets, MetaIndexParser);
+       // if we get a timeout if fail
+       if(LookupTag(Message,"FailReason") == "Timeout" || 
+	  LookupTag(Message,"FailReason") == "TmpResolveFailure") {
+	 Item::Failed(Message,Cnf);
+	 return;
+       }
+
+       // queue a pkgAcqMetaIndex with no sigfile
+       new pkgAcqMetaIndex(Owner, MetaIndexURI, MetaIndexURIDesc,
+			   MetaIndexShortDesc,
+			   "", IndexTargets, MetaIndexParser);
+     }
 
    if (Cnf->LocalOnly == true || 
        StringToBool(LookupTag(Message,"Transient-Failure"),false) == false)
-   {      
-      // Ignore this
+     {      
+       // Ignore this
       Status = StatDone;
       Complete = false;
       Dequeue();
       return;
-   }
+     }
    
    Item::Failed(Message,Cnf);
 }
@@ -690,13 +720,44 @@ bool pkgAcqMetaIndex::VerifyVendor()
 void pkgAcqMetaIndex::Failed(string Message,pkgAcquire::MethodConfig *Cnf)
 {
    if (AuthPass == true)
-   {
-      // gpgv method failed
-      _error->Warning("GPG error: %s: %s",
-                      Desc.Description.c_str(),
-                      LookupTag(Message,"Message").c_str());
-   }
+     {
+       // gpgv method failed
+       _error->Warning("GPG error: %s: %s",
+		       Desc.Description.c_str(),
+		       LookupTag(Message,"Message").c_str());
+     }
+   else
+     {
+       // If this is a transient failure, we use the old version, if we
+       // have one.
+       //
+       if (StringToBool(LookupTag(Message,"Transient-Failure"),false))
+	 {
+	   string Final = _config->FindDir("Dir::State::lists");
+	   Final += URItoFileName(RealURI);
+	   struct stat Buf;
+	   if (stat(Final.c_str(),&Buf) == 0)
+	     {
+	       DestFile = Final;
 
+	       if (SigFile != "")
+		 {
+		   // There was a signature file, so pass it to gpgv
+		   // for verification
+
+		   if (_config->FindB("Debug::pkgAcquire::Auth", false))
+		     std::cerr << "Metaindex acquired, queueing gpg verification ("
+			       << SigFile << "," << DestFile << ")\n";
+		   AuthPass = true;
+		   Desc.URI = "gpgv:" + SigFile;
+		   QueueURI(Desc);
+		   Mode = "gpgv";
+		   return;
+		 }
+	     }
+	 }
+     }
+       
    // No Release file was present, or verification failed, so fall
    // back to queueing Packages files without verification
    QueueIndexes(false);
